Changed header #ifndef's to ATEN_*. Custom CLI option handling - replaced master::prepare_cli() with master::cli_debug(). Changed all colour definitions throughout to GLfloat[4]. Removed canvas::swap_buffers() [empty]. New prefs window. Fixed instance of TColourFrame to update properly.

+++ src/render/objects.cpp
@@ -261,7 +261,7 @@ void canvas::gl_sphere(double radius, bool filled)
 	// Don't use this to render objects to the view - create a display list first!
 	int i, j;
 	int lats = prefs.get_atom_detail();
-	int longs = lats * 1.5;
+	int longs = int(lats * 1.5);
 	double lat0, z0, zr0, lat1, z1, zr1, lng, x, y;
 	glPolygonMode(GL_FRONT_AND_BACK, (filled ? GL_FILL : GL_LINE));
 	for(i = 0; i <= lats; i++)
